# Horizontal Task Fusion for Metrics Computation